rename module to `on_disk_text_index`
