+++ lib/segment/src/index/field_index/full_text_index/mod.rs
@@ -3,7 +3,7 @@ use std::path::PathBuf;
 use common::universal_io::MmapFile;
 
 use self::immutable_text_index::ImmutableFullTextIndex;
-use self::mmap_text_index::OnDiskFullTextIndex;
+use self::on_disk_text_index::OnDiskFullTextIndex;
 use self::mutable_text_index::MutableFullTextIndex;
 use crate::data_types::index::TextIndexParams;
 
@@ -11,7 +11,7 @@ pub mod full_text_index_read;
 mod immutable_text_index;
 mod inverted_index;
 mod lifecycle;
-pub mod mmap_text_index;
+pub mod on_disk_text_index;
 mod mutable_text_index;
 pub mod read_only;
 mod read_ops;
