+++ lib/segment/src/index/field_index/full_text_index/on_disk_text_index/lifecycle.rs
